backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
firebase-admin==6.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
httpx==0.28.1
bcrypt==4.2.1
